fix dnf-system-upgrade.service
The Software Updates spec recommends using `OnFailure=reboot.target` to
reboot if the unit fails. In a discussion on systemd-devel, they
recommended using `FailureAction=reboot` instead, so I changed that.

But: it turns out that `OnFailure` goes in the `[Unit]` section, but
`FailureAction` is a `[Service]` option. So systemd would refuse to load
our service with the message:

    Unknown lvalue 'FailureAction' in section 'Unit'

...and then either `packagekit-offline-update.service` reboots the
system or the system hangs. Whoops.