Revert 'do-upgrade' back to 'upgrade', add unit test
This was a simple stylistic change but since we use the action name to
choose the functions to run, `dnf system-upgrade ... do-upgrade` failed
to find the `upgrade` functions and exited silently rather than actually
upgrading. Whoops.

This fixes that, and adds a unit test to make sure that doesn't happen
in the future.

This plus the previous commit (e9f83cc) should fix #4.