.service: Drop StandardOutput=journal+console
The Fedora packaging guidelines suggest that services shouldn't override
the default StandardOutput setting, but we want status messages on the
console.

But: it turns out that plymouth prints the messages that we send it to
the terminal, so having DNF output onto the console means that we get
duplicate status messages for each event.

Drop this line and use the default setting; we'll still log all the
information to the systemd journal, and we'll still get status messages
on the plymouth console.

This resolves #13.