bump version number to 0.7.0
As mentioned in commit c9ac595, I've branched the code. f21-branch will
be version 0.5.x, and it will maintain compatibility with DNF < 1.1.0.

From here on out, master will require DNF 1.1.0 or later (currently
that's F22 through F24). So we need to bump the version number.

I *was* going to use 0.6.x, but I didn't want to confuse things with DNF
0.6.x, so we're going straight to 0.7.0. Whee!