fedup.sh: add `reboot` and other passthrough options
People are pretty used to 'fedup'. We might want to (grudgingly) allow
it to be used as the primary interface to the system-upgrade plugin, at
least for a while.

So: it should be capable of handling all the stuff that we'll need to
make upgrades work. So we need to add the `reboot` command, and a few of
the commonly-used DNF flags.

This commit adds `reboot`, `--best`, `--allowerasing`, and
`--disableplugin` (which DNF supports now, I guess).

It also adds unit tests to make sure those do what we expect.