plugin: hide graphical splash during upgrade with old DNF
Since DNF < 1.0.1 can't talk to plymouth during the upgrade, we should
just hide the graphical splash and let the user see the DNF progress
output on the console.

This patch makes the plugin tell plymouth to hide the splash screen if
DNF is older than 1.0.1.

NOTE: halfline tells me that 'hide-splash' causes plymouthd to detach
from the tty; hopefully this does not have unintended side-effects.