plugin: use unicode_literals
Do like the rest of DNF does: 'from __future__ import unicode_literals'.

This makes all literal strings unicode objects in Python 2.x, which is a
little strange but that's what the rest of DNF expects.

Plugins that fail to do so can cause problems in DNF; see rhbz#1262082.

This should resolve #17.