spec: clean old fedup junk and migrate pkg cache
If we're replacing fedup, then we should clean up old fedup junk - boot
images, bootloader entry, /system-upgrade symlink, etc. - but we should
try to keep the package cache, if possible, so the user doesn't have to
download everything again.

The easiest way to handle this is to do it in %pre, so we can move the
cache and then just tell fedup to clean up whatever's left *before* it
gets deleted.

Resolves: rhbz#1264948