plugin: propagate --best/--allowerasing to upgrade (rhbz#1266589)
For the upgrade to work correctly, we need to do the depsolving the same
way during "upgrade" as we do during "download".

Since the DNF flags --best and --allowerasing change how the depsolving is
done, we need to carry their state into the upgrade.

So: save them into the State after a successful test transaction, then
restore them at the start of the upgrade.

(Also update the tests to check this)