fedup.sh: fix `fedup --network VERSION`
Because of how dnf handles --releasever, we can't reset it from inside
the plugin (see https://bugzilla.redhat.com/show_bug.cgi?id=1212341).

Rather than just failing to do what the user asked for, we need to
actually exit with an error if we get to the point where we're trying to
reset releasever.

To avoid this happening, we add a little check to fedup.sh that rewrites
`--network VERSION` to `download --releasever VERSION`, which allows the
old `fedup --network VERSION` form of the command to continue working.