plugin: fix traceback in run_upgrade
Argh - F21-era python will trace back if you try to compare a
StrictVersion to a unicode object. (A string is fine, though..)

So when we try to run the upgrade, our DNF version check causes a
traceback and the upgrade crashes and reboots. Whoops.

Fix this by comparing two StrictVersion objects, and add a test case to
make sure that comparison works as expected.