plymouth: do "hide-message" after each message
Some plymouth plugins toss the current message if you display-message a
new one, but... not all of them. Which can cause weird problems.

So! We need to explicitly "hide-message" the previous message before
displaying a new message, apparently.

This makes the Plymouth object automatically handle that for us on each
call to .message().

Related: rhbz#1267949