Disable console blanking during upgrade
Blanking is disabled on the foreground terminal (/dev/tty0) after
upgrade is started. There doesn't seem to be a way to query current
settings, so when dnf exits, blanking is set to 10 minutes, which
should be good enough not to burn anyone's screen (especially that we
reboot soon after, so we probably could get away with leaving it
disabled). This applies only to the foreground tty when upgrade is
started, so if the user switches virtual terminals, still only one
will have blanking disabled.

Fixes #26.