README: add link to the bug tracker
The text in README.md is a bit different than the one in the man page,
on purpose: people who are reading the man page will do so on the
installed system, and should use the bugtracker. People who are
reading README.md will often do that on github, so the text should
be more generic and should also encourage the use of github issues.

Fixes #36.