plugin: fix crash at startup with zh_CN.UTF-8
python2 does not like it when we try to interpolate non-ascii strings
returned by _() into a unicode object (remember, our string literals are
unicode objects in py2 because we're using unicode_literals).

Make sure the values returned by _() here are converted to unicode(ish)
before doing interpolation and we're good.

Closes #32, closes #34
Resolves: rhbz#1278031