Always remove /system-update symlink on failure
If something prevents the plugin from loading, the symlink would
not be removed, leading to reboot loops. Add a duplicate mechanism
to remove the symlink as a safety feature.

Note that the unit is conditionalized on
/system-update/.dnf-system-upgrade existing. It is not strictly
the same condition that is used to detect if it is "our" upgrade,
but it is close enough.

Fixes #37.