Send DNF stdout/stderr to text console (#1317195)
We previously dropped this, because it was causing duplicate text on
systems with plymouth - see #13 - but the side-effect is that now
there's no output whatsoever on systems *without* plymouth.

Too much output is better than not enough. Restore the output from DNF
here, and we'll try to fix plymouth so it doesn't duplicate its output
to its "details" view.

This reverts commit dcc334939aa888ae8e15291a16982062ec446226.
Resolves: RHBZ#1317195